tools/requirements.txt
gwpy
pint-pulsar
astropy
numpy
scipy
matplotlib
